pyproject.toml
[tool.poetry]
name = "online-store-django"
version = "0.1.0"
description = ""
authors = ["egor <sharapov.egor00@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"
ipython = "^8.22.2"
pytils = "^0.4.1"
django-crispy-forms = "^2.1"
crispy-bootstrap4 = "^2024.1"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
